Revert "switch plugin style from inheritance to a role"
This reverts commit 67ce53daf10ca728b61469f98f2a081737e0a7fe.

I did not sufficiently appreciate the effect it would have on the order
of method construction and resolution and how difficult it would make
to wrap the plugin-supplied uri_for_action method.

This was not helped by the fact that the plugin went out of its way
to delay its application as a method modifier either. There were good
reasons for doing this, all due to the fact that it allowed closing over
the cache hash. This obviated the need to find somewhere on some object
to store it, the need to go looking for it there every time, and also
the need to forever check whether to fall back to the original method
while the cache was not yet present even though this is only possible
briefly early on during Catalyst startup.

Unfortunately this all made class composition ordering too complicated.
It turns out that a multiple-inheritance-based plugin is easier to deal
with – which is not a punchline I would have expected.